requirements.txt
absl-py==2.1.0
cachetools==5.3.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.3.2
cycler==0.11.0
fonttools==4.38.0
google-auth==2.27.0
google-auth-oauthlib==0.4.6
grpcio==1.60.0
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.4
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.2
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
